fraud-service/requirements.txt
uvicorn
fastapi
pydantic
httpx
python-dotenv
